Move tests into their own test class
