+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/ivy/IvyDynamicRevisionRemoteResolveIntegrationTest.groovy
@@ -108,263 +108,6 @@ dependencies {
         checkResolve "group:projectA:1.+": "group:projectA:1.2", "group:projectB:latest.integration": "group:projectB:2.2"
     }
 
-    def "can use a custom metadata provider"() {
-        given:
-        buildFile << """
-          repositories {
-              ivy {
-                  name 'repo'
-                  url '${ivyHttpRepo.uri}'
-                  metadataSupplier(MP)
-              }
-          }
-          
-            configurations { compile }
-            if (project.hasProperty('refreshDynamicVersions')) {
-                configurations.all {
-                    resolutionStrategy.cacheDynamicVersionsFor 0, "seconds"
-                }
-            }
-            dependencies {
-                compile group: "group", name: "projectA", version: "1.+"
-                compile group: "group", name: "projectB", version: "latest.release"
-            }
-
-          import javax.inject.Inject
-     
-          class MP implements ComponentMetadataSupplier {
-          
-            final RepositoryResourceAccessor repositoryResourceAccessor
-            
-            @Inject
-            MP(RepositoryResourceAccessor accessor) { repositoryResourceAccessor = accessor }
-          
-            int count
-          
-            void supply(ComponentMetadataSupplierDetails details) {
-                def id = details.id
-                println "Providing metadata for \$id"
-                repositoryResourceAccessor.withResource("\${id.group}/\${id.module}/\${id.version}/status.txt") {
-                    details.result.status = new String(it.bytes)
-                }
-                println "Metadata rule call count: \${++count}"
-            }
-          }
-"""
-        when:
-        def projectA1 = ivyHttpRepo.module("group", "projectA", "1.1").publish()
-        def projectA2 = ivyHttpRepo.module("group", "projectA", "1.2").publish()
-        def projectB1 = ivyHttpRepo.module("group", "projectB", "1.1").publish()
-        def projectB2 = ivyHttpRepo.module("group", "projectB", "2.2").publish()
-        ivyHttpRepo.module("group", "projectA", "2.0").publish()
-
-
-        and:
-        expectGetStatusOf(projectB1, 'release')
-        expectGetStatusOf(projectB2, 'integration')
-        expectGetDynamicRevision(projectA2)
-        expectGetDynamicRevision(projectB1)
-
-        then: "custom metadata rule prevented parsing of ivy descriptor"
-        checkResolve "group:projectA:1.+": "group:projectA:1.2", "group:projectB:latest.release": "group:projectB:1.1"
-        outputContains 'Providing metadata for group:projectB:2.2'
-        outputContains 'Providing metadata for group:projectB:1.1'
-        !output.contains('Providing metadata for group:projectA:1.1')
-
-        and: "creates a new instance of rule each time"
-        !output.contains('Metadata rule call count: 2')
-    }
-
-    def "handles errors in a custom metadata provider"() {
-        given:
-        buildFile << """
-          repositories {
-              ivy {
-                  name 'repo'
-                  url '${ivyHttpRepo.uri}'
-                  metadataSupplier(MP)
-              }
-          }
-          
-            configurations { compile }
-            
-            dependencies {
-                compile group: "group", name: "projectA", version: "1.+"
-                compile group: "group", name: "projectB", version: "latest.release"
-            }
-
-          class MP implements ComponentMetadataSupplier {
-          
-            void supply(ComponentMetadataSupplierDetails details) {
-                throw new NullPointerException("meh: error from custom rule")
-            }
-          }
-"""
-        def projectA1 = ivyHttpRepo.module("group", "projectA", "1.1").publish()
-        def projectB1 = ivyHttpRepo.module("group", "projectB", "1.1").publish()
-
-        when:
-        expectListVersions(projectA1)
-        projectA1.ivy.expectGet()
-        expectListVersions(projectB1)
-
-        fails 'checkDeps'
-
-        then:
-
-        errorOutput.contains('Could not resolve group:projectB:latest.release')
-        failure.assertHasCause('meh: error from custom rule')
-    }
-
-    def "custom metadata provider doesn't have to do something"() {
-        given:
-        buildFile << """
-          repositories {
-              ivy {
-                  name 'repo'
-                  url '${ivyHttpRepo.uri}'
-                  metadataSupplier(MP)
-              }
-          }
-          
-            configurations { compile }
-            
-            dependencies {
-                compile group: "group", name: "projectA", version: "1.+"
-                compile group: "group", name: "projectB", version: "latest.integration"
-            }
-
-          class MP implements ComponentMetadataSupplier {
-          
-            void supply(ComponentMetadataSupplierDetails details) {
-                // does nothing
-            }
-          }
-"""
-        def projectA1 = ivyHttpRepo.module("group", "projectA", "1.1").publish()
-        def projectB1 = ivyHttpRepo.module("group", "projectB", "1.1").publish()
-
-        when:
-        expectGetDynamicRevision(projectA1)
-        expectGetDynamicRevision(projectB1)
-
-        then:
-        checkResolve "group:projectA:1.+": "group:projectA:1.1",
-            "group:projectB:latest.integration": "group:projectB:1.1"
-    }
-
-    def "can use a single remote request to get status of multiple components"() {
-        given:
-        buildFile << """
-          repositories {
-              ivy {
-                  name 'repo'
-                  url '${ivyHttpRepo.uri}'
-                  metadataSupplier(MP)
-              }
-          }
-          
-            configurations { compile }
-            if (project.hasProperty('refreshDynamicVersions')) {
-                configurations.all {
-                    resolutionStrategy.cacheDynamicVersionsFor 0, "seconds"
-                }
-            }
-            dependencies {
-                compile group: "group", name: "projectA", version: "1.+"
-                compile group: "group", name: "projectB", version: "latest.release"
-            }
-
-          import javax.inject.Inject
-     
-          class MP implements ComponentMetadataSupplier {
-          
-            final RepositoryResourceAccessor repositoryResourceAccessor
-            
-            @Inject
-            MP(RepositoryResourceAccessor accessor) { repositoryResourceAccessor = accessor }
-            
-            int calls
-            Map<String, String> status = [:]
-          
-            void supply(ComponentMetadataSupplierDetails details) {
-                def id = details.id
-                println "Providing metadata for \$id"
-                repositoryResourceAccessor.withResource("status.txt") {
-                    if (status.isEmpty()) {
-                        println "Parsing status file call count: \${++calls}"
-                        it.withReader { reader ->
-                            reader.eachLine { line ->
-                                if (line) {
-                                   def (module, st) = line.split(';')
-                                   status[module] = st
-                                }
-                            }
-                        }
-                        println status
-                    }
-                }
-                details.result.status = status[id.toString()]
-            }
-          }
-"""
-        when:
-        def projectA1 = ivyHttpRepo.module("group", "projectA", "1.1").publish()
-        def projectA2 = ivyHttpRepo.module("group", "projectA", "1.2").publish()
-        def projectB1 = ivyHttpRepo.module("group", "projectB", "1.1").publish()
-        def projectB2 = ivyHttpRepo.module("group", "projectB", "2.2").publish()
-        ivyHttpRepo.module("group", "projectA", "2.0").publish()
-
-
-        and:
-        def statusFile = temporaryFolder.createFile("versions.status")
-        statusFile << '''group:projectA:1.1;release
-group:projectA:1.2;release
-group:projectB:1.1;release
-group:projectB:2.2;integration
-'''
-        server.expectGet("/repo/status.txt", statusFile)
-        expectGetDynamicRevision(projectA2)
-        expectGetDynamicRevision(projectB1)
-
-        then: "custom metadata rule prevented parsing of ivy descriptor"
-        checkResolve "group:projectA:1.+": "group:projectA:1.2", "group:projectB:latest.release": "group:projectB:1.1"
-        outputContains 'Providing metadata for group:projectB:2.2'
-        outputContains 'Providing metadata for group:projectB:1.1'
-        !output.contains('Providing metadata for group:projectA:1.1')
-
-        and: "remote status file parsed only once"
-        outputContains 'Parsing status file call count: 1'
-        !output.contains('Parsing status file call count: 2')
-
-        when: "resolving the same dependencies"
-        server.expectHead("/repo/status.txt", statusFile)
-        server.expectHead("/repo/status.txt", statusFile)
-        checkResolve "group:projectA:1.+": "group:projectA:1.2", "group:projectB:latest.release": "group:projectB:1.1"
-
-        then: "should parse the result from cache"
-        output.contains('Parsing status file call count: 1')
-
-        when: "force refresh dependencies"
-        executer.withArgument("-PrefreshDynamicVersions")
-        statusFile.text = '''group:projectA:1.1;release
-group:projectA:1.2;release
-group:projectB:1.1;release
-group:projectB:2.2;release
-'''
-        server.expectHead("/repo/status.txt", statusFile)
-        server.expectGet("/repo/status.txt", statusFile)
-        expectListVersions(projectA2)
-        expectGetDynamicRevision(projectB2)
-
-        then: "shouldn't use the cached resource"
-        checkResolve "group:projectA:1.+": "group:projectA:1.2", "group:projectB:latest.release": "group:projectB:2.2"
-        outputContains 'Providing metadata for group:projectB:2.2'
-        !output.contains('Providing metadata for group:projectB:1.1')
-        !output.contains('Providing metadata for group:projectA:1.1')
-
-    }
-
     @Unroll
     def "uses latest version from version range with #identifier characters"() {
         given:
